feat: default input value
