+++ packages/combobox/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./ComboBox"
+export * from "./ComboBoxItem"
