--- packages/combobox/src/react-app-env.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="react-scripts" />
-declare module "*.png" {
-  const value: any
-  export default value
-}
-declare module "*.gif" {
-  const value: any
-  export default value
-}
